Polish goose2 settings layout
