+++ ui/goose2/src/features/chat/ui/__tests__/ChatInput.test.tsx
@@ -315,7 +315,7 @@ describe("ChatInput", () => {
     expect(dispatchEventSpy).toHaveBeenCalledWith(
       expect.objectContaining({
         type: OPEN_SETTINGS_EVENT,
-        detail: { section: "compaction" },
+        detail: { section: "general" },
       }),
     );
