+++ ui/goose2/src/features/settings/ui/SettingsView.tsx
@@ -1,12 +1,9 @@
-import { AppearanceSettings } from "./AppearanceSettings";
 import { DoctorSettings } from "./DoctorSettings";
 import { ProvidersSettings } from "./ProvidersSettings";
 import { VoiceInputSettings } from "./VoiceInputSettings";
 import { GeneralSettings } from "./GeneralSettings";
-import { CompactionSettings } from "./CompactionSettings";
 import { ProjectsSettings } from "./ProjectsSettings";
 import { ChatsSettings } from "./ChatsSettings";
-import { AboutSettings } from "./AboutSettings";
 import type { SectionId } from "./settingsSections";
 import { PageShell } from "@/shared/ui/page-shell";
 
@@ -16,16 +13,13 @@ interface SettingsViewProps {
 
 export function SettingsView({ activeSection }: SettingsViewProps) {
   return (
-    <PageShell contentClassName="gap-0">
-      {activeSection === "appearance" && <AppearanceSettings />}
+    <PageShell contentWidth="narrow" contentClassName="gap-0">
       {activeSection === "providers" && <ProvidersSettings />}
-      {activeSection === "compaction" && <CompactionSettings />}
       {activeSection === "voice" && <VoiceInputSettings />}
       {activeSection === "doctor" && <DoctorSettings />}
       {activeSection === "general" && <GeneralSettings />}
       {activeSection === "projects" && <ProjectsSettings />}
       {activeSection === "chats" && <ChatsSettings />}
-      {activeSection === "about" && <AboutSettings />}
     </PageShell>
   );
 }
